Move MappingEntity and Container to data-model
