+++ data-model/src/main/java/org/pdxfinder/admin/pojos/MappingContainer.java
@@ -1,6 +1,5 @@
-package org.pdxfinder.admin.models;
+package org.pdxfinder.admin.pojos;
 
-import java.util.Map;
 import java.util.TreeMap;
 
 /*
