+++ data-model/src/main/java/org/pdxfinder/admin/pojos/MappingEntity.java
@@ -1,4 +1,4 @@
-package org.pdxfinder.admin.models;
+package org.pdxfinder.admin.pojos;
 
 import java.util.ArrayList;
 import java.util.Date;
